+++ ForestManagement/functions/schema/Invoke-FMSchema.ps1
@@ -12,6 +12,10 @@
 			Use the module's configuration settings to govern schema admin credentials.
 			The configuration can be read with Get-PSFConfig and updated with Set-PSFConfig.
 		
+		.PARAMETER InputObject
+			Test results provided by the associated test command.
+			Only the provided changes will be executed, unless none were specified, in which ALL pending changes will be executed.
+		
 		.PARAMETER Server
 			The server / domain to work with.
 		
@@ -36,6 +40,9 @@
 	
 	[CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'Low')]
 	Param (
+		[Parameter(ValueFromPipeline = $true)]
+		$InputObject,
+		
 		[PSFComputer]
 		$Server,
 
@@ -78,8 +85,6 @@
 		if (Test-PSFFunctionInterrupt) { return }
 		#endregion Resolve Credentials
 
-		$testResult = Test-FMSchema @parameters
-
 		# Prepare parameters to use for when discarding the schema credentials
 		if ($cred -and ($cred -ne $Credential)) { $removeParameters['SchemaAccountCredential'] = $cred }
 	}
@@ -87,13 +92,17 @@
 	{
 		if (Test-PSFFunctionInterrupt) { return }
 
-		$testResultsSorted = $testResult | Sort-Object {
+		if (-not $InputObject) {
+			$InputObject = Test-FMSchema @parameters
+		}
+
+		$testResultsSorted = $InputObject | Sort-Object {
 			if ($_.Type -eq 'Decommission') { 0 }
 			elseif ($_.Type -eq 'Rename') { 2 }
 			elseif ($_.Type -eq 'ConfigurationOnly') { 3 }
 			else { 1 }
 		}
-		:main foreach ($testItem in ($testResultsSorted)) {
+		:main foreach ($testItem in $testResultsSorted) {
 			switch ($testItem.Type) {
 				#region Create new Schema Attribute
 				'ConfigurationOnly' {